+++ Model/Content.php
@@ -5,13 +5,29 @@ namespace LinkCMS\Model;
 use LinkCMS\Actor\Core;
 
 abstract class Content {
-    var $author;
     var $draftContent;
-    var $draftTimestamp;
+    var $draftModifiedDate;
+    var $id;
+    var $pubDate;
     var $publishedContent;
-    var $publishedTimestamp;
+    var $publishedModifiedDate;
     var $slug;
-    var $taxonomy;
+    var $status;
     var $title;
-    static $type;
+    var $type;
+
+    public function __construct($data=false) {
+        if ($data) {
+            if (is_array($data)) {
+                foreach ($data as $param => $value) {
+                    if ($value == 'null' || $value == 'false')
+                        continue;
+                    if (($param == 'publishedContent' || $param == 'draftContent') && is_string($value)) {
+                        $value = json_decode($value);
+                    }
+                    $this->{$param} = $value;
+                }
+            }
+        }   
+    }
 }