+++ Public/js/lib/hat.js
@@ -82,7 +82,23 @@
                       Internal.sorting.option('sorting', true);
                   }
               });
-          } 
+          },
+          removeBlock: function(block, defaultAction=true) {
+            var blockId = block.el.id;
+            if (Blocks.hasOwnProperty(blockId)) {
+                block.getPosition();
+                if (defaultAction) {
+                    let newFocus = (block.position.first) ? block.el.nextSibling.id : block.el.previousSibling.id;
+                    Blocks[newFocus].focus();
+                }
+                block.el.remove();
+                Blocks.splice(blockId);
+                BlockCount--;
+                if (defaultAction) {
+                    Events.fire('blockChanged');
+                }
+            }
+          }
       };
       let Interface = {
           addBlock: function(focus=true,position=false, type=window.Hat.getDefault(), data=false) {
@@ -130,23 +146,27 @@
                       contents.settings.id = key;
                   }
                   content.push(contents);
-              }            return content;
+              }            
+              return content;
           },
           loadBlock: function(data) {
               let block = Interface.addBlock(false, false, data.type, data);
           },
-          removeBlock: function(block) {
-              var blockId = block.el.id;
+          removeAllBlocks: function(replacementData) {
+            for (let [id, block] of Object.entries(Blocks)) {
+                Internal.removeBlock(block, false);
+            };
+            if (replacementData) {
+                replacementData.forEach(function(blockData) {
+                    Interface.loadBlock(blockData);
+                });
+            } else {
+                Interface.addBlock();
+            }
+          },
+          removeBlock: function(block, force=false) {
               if (BlockCount > 1){
-                  if (Blocks.hasOwnProperty(blockId)) {
-                      block.getPosition();
-                      let newFocus = (block.position.first) ? block.el.nextSibling.id : block.el.previousSibling.id;
-                      Blocks[newFocus].focus();
-                      block.el.remove();
-                      Blocks.splice(blockId);
-                      BlockCount--;
-                      Events.fire('blockChanged');
-                  }
+                Internal.removeBlock(block);
               }
           },
       };
@@ -957,6 +977,9 @@
                   }
               }, 1);
           });
+          toolbar.contextButtons.forEach(function(el) {
+            el.setAttribute('disabled', 'disabled');
+        });
       }
 
       addFormattingButtons() {
@@ -1038,6 +1061,7 @@
                   return true;
               }
               new SelectionWrapper('a', toolbar.parentBlock.view, values);
+              toolbar.checkFormatting();
           });
           link.modalContainer.addEventListener('canceled', function(e) {
               toolbar.returnCursor(sel, range);
@@ -1083,16 +1107,17 @@
           this.contextButtons.push(el);
           toolbar.container.append(el);
       }
-
+      
       addUnlinkButton() {
-          let toolbar = this;
-          this.unlinkBtn = new DomButton('Unlink text', 'unlink');
-          this.unlinkBtn.addEventListener('click', function(e) {
-              e.preventDefault();
-              toolbar.unlink();
-          });
-          toolbar.container.append(this.unlinkBtn);
-      }
+        let toolbar = this;
+        this.unlinkBtn = new DomButton('Unlink text', 'unlink');
+        this.unlinkBtn.setAttribute('disabled', true);
+        this.unlinkBtn.addEventListener('click', function(e) {
+            e.preventDefault();
+            toolbar.unlink();
+        });
+        toolbar.container.append(this.unlinkBtn);
+    }
 
       checkForDeepTag(tag) {
           if (!sel) {
@@ -1114,23 +1139,26 @@
           let found = false;
           let sel = window.getSelection();
           let range = false;
+          let anchor = false;
           if (sel && sel.rangeCount > 0) {
               range = sel.getRangeAt(0);
           }
-          let anchor = sel.anchorNode.parentElement;
-          let focus = sel.focusNode.parentElement;
-          // Check that we're actually in the edit container
-          if (this.parentBlock.editEl.contains(anchor) && ( anchor.tagName.toLowerCase() == tag || focus.tagName.toLowerCase() == tag)) {
-              found = true;
+          if (sel.anchorNode) {
+            anchor = sel.anchorNode.parentElement;
+            let focus = sel.focusNode.parentElement;
+            // Check that we're actually in the edit container
+            if (this.parentBlock.editEl.contains(anchor) && ( anchor.tagName.toLowerCase() == tag || focus.tagName.toLowerCase() == tag)) {
+                found = true;
+            }
           }
           switch (tag) {
-              case 'a':
-                  if (found || this.parentBlock.editEl.contains(anchor) && this.checkForDeepTag(tag)) {
-                      btn.removeAttribute('disabled');
-                  } else {
-                      btn.setAttribute('disabled', true);
-                  }
-                  break;
+            case 'a':
+                if (found || (anchor && this.parentBlock.editEl.contains(anchor)) && this.checkForDeepTag(tag)) {
+                    btn.removeAttribute('disabled');
+                } else {
+                    btn.setAttribute('disabled', true);
+                }
+                break;
           }
       }
 
@@ -1154,6 +1182,9 @@
           this.parentBlock.editEl.addEventListener('keydown', debounce((e) => {
               toolbar.checkFormatting();
           }, 350));
+          this.parentBlock.editEl.addEventListener('click', function(e) {
+            toolbar.checkFormatting();
+          });
           this.parentBlock.editEl.addEventListener('focusin', function(e) {
               toolbar.checkFormatting();
           });
@@ -1413,7 +1444,9 @@
       };
       let Interface = {   
           createEditor: function(el) {
-              EditorRegistry.add(new Editor(el));
+            let ed = new Editor(el);  
+            EditorRegistry.add(ed);
+            return ed;
           },
           getBlock: function(blockName) {
               if (Interface.hasBlock(blockName)) {
@@ -1460,15 +1493,23 @@
               BlockRegistry.objects[slug] = blockObj;
           },
           start: function(options) {
-              for (let [key, value] of Object.entries(options)) {
-                  Options[key] = value;
-              }            if (Options.data) {
-                  EditorRegistry.add(new Editor(document.querySelector(Options.selector), Options.data));
-              } else if (Options.init) {
-                  for (var el of document.querySelectorAll(Options.selector)) {
-                      Interface.createEditor(el);
-                  }
-              }
+            for (let [key, value] of Object.entries(options)) {
+                Options[key] = value;
+            } 
+            if (Options.data) {
+                let ed = new Editor(document.querySelector(Options.selector), Options.data);
+                EditorRegistry.add(ed);
+                return ed;
+            } else if (Options.init) {
+                let elements = document.querySelectorAll(Options.selector);
+                if (elements.length == 1) {
+                    return Interface.createEditor(elements[0]);
+                } else {
+                    for (var el of elements) {
+                        Interface.createEditor(el);
+                    }
+                }
+            }
           }
       };
       return Interface;